[Tidy up] Moved test into fields integration test folder
